Updates subtxt with formatting and config options

+++ subtxt.sh
@@ -1,26 +1,34 @@
 #!/bin/sh
-subs="following.txt"
+# subtxt - pulls and prints twtxt files
+CONFIG="/home/example-user/.subtxt"
+SUBS="$CONFIG/following.txt"
+FEEDS="$CONFIG/feeds"
+TIMEOUT=10
 
-mkdir -p feeds
+# Make dirs if they don't exist, clean up old main feed
+mkdir -p "$FEEDS"
+touch "$SUBS"
+if [ -e main.feed ]; then rm main.feed; fi
 
 # Fetch all feeds
-touch main.feed
-cat "$subs" | sed -e "/^[\s\#]/d" | while read l; do
+sed -e "/^[\s\#]/d" "$SUBS" | while read -r l; do
 	user=$(echo "$l" | cut -d"	" -f1)
 	url=$(echo "$l" | cut -d"	" -f2-)
+	feed="$FEEDS"/"$user".feed
 
-	feed=feeds/"$user".feed
-	# Multi-step to handle feeds feeds
-	curl -s "$url" -o "$feed"
+	# Multi-step to handle local feeds
+	curl -s "$url" -o "$feed" --max-time $TIMEOUT
 
-	# Remove all settings and lines that start with whitespace
-	sed "/^\[\s\#]/d" "$feed" > "$feed".tmp
-
-	# Prepend posts with nicknames
-	sed -e "s/^/$user	/" "$feed".tmp >> main.feed
-	rm "$feed".tmp
+	# Check to see if curl failed
+	if ! curl -s "$url" -o "$feed" --max-time $TIMEOUT; then
+		echo "ERROR: Couldn't fetch $user ($url)\n"
+	else
+		# Remove all settings and lines that start with whitespace
+		sed "/^[\s\#]/d" "$feed" > "$feed".tmp
+		sed -e "s/^/$user\t/" "$feed".tmp >> main.feed
+	fi
 done
 
 # cat all feeds, sorted by datetime
-sort -k 2,2r main.feed
-rm main.feed
+sort -k 2,2r main.feed > tmp && mv tmp main.feed
+column -s"	" -t main.feed
